Change migration slightly for reducing risk during deployment
Update the data on most recent ARs first so that anything active in the queue get's updated first. For when queue/app is in use during deployment.

cc @krsnachandra

Hmmm.. I wonder if we should do a check to make sure all ARs have this col set b/c during the deployment it is possible that some AR records still dont have this column set due to phased (`blue-green`) deployment. Will talk with @jmatsuba about this.