Production has many inactive locations. Fix noise in selector
The selector at the top of the assistance queue would show 3 inactive locations on production (not noticed in development env).

I added a boolean `active` to Location model and use it to only filter and show `active` locations (ie dont include Kelowna, London and Halifax by setting them to inactive via the migration).

We can activate these locations in the queue later by setting `active=true` for those in the future. This can only be done via `rails c` on production, but that's ok for now.

Unforuntately this new feature had to be added in this enhancement.

cc @krsnachandra & @jmatsuba