Admin: Enrol in cohort: Fixed dev seeds
Add 2 upcoming cohorts so that there are some that are selectable. Otherwise every dev that needs to test this would have to manually create upcoming cohorts.

Also, noticed that we use faker for a bunch of unique fields like github_username and email, but don't use Faker's `unique` method. Fixing as part of this.