Resolves #668
If in a workbook, an activity will only appear as stretch if the "item" (in the workbook _config.yml) is marked as `stretch: true`.

Also refactored how presenters work while doing this.

They now take in an options hash, allowing me to pass in a (optional) workbook instance into ActivityPresenter. This cleans up some of the interface I dirtied on ActivityPresenter

cc @meech-ward