resolves #672
No longer 500 when location AND keywords exist

I had to hack around a weird behavior from `pg_search` when dealing with the name of the joined table. The expected name would be the actual name of the table (`users`) but when using `pg_search` AR's `joins` and references don't work as expected.