re-rejig Program dropdown for lectures & activities
Search divider based reorg of the Program dropdown wasn't looking as good as I had pictured in my head when I suggested it.

Also it had too many dividers if teachers or even for alumni.

Removed mention of search and just called the "Activities" and "Lectures" and moved them higher up.

cc @krsnachandra