Support lack of prep content
If there isn't a prep activity, which perhaps could happen with some programs, then redirect back to welcome page with a flash alert

This is not the best UX and only a temporary solution to avoid a 500 error on /prep. We will need to revisit the workflow here more holistically later on.